Fixed comments given during the meeting
