--- components/org.wso2.configuration.diff.creater/src/main/java/org/wso2/configuration/diff/creater/ConfigLoader.java
@@ -1,169 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-package org.wso2.configuration.diff.creater;
-
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.wso2.configuration.diff.creater.exception.ConfigMigrateException;
-import org.wso2.configuration.diff.creater.utils.MigrationConstants;
-
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-/**
- * Class for loading config files.
- */
-public class ConfigLoader {
-
-    private Map<String, File> defaultXMLFiles = new HashMap<>();
-    private Map<String, File> migratedXMLFiles = new HashMap<>();
-    private Map<String, File> defaultPropertiesFiles = new HashMap<>();
-    private Map<String, File> migratedPropertiesFiles = new HashMap<>();
-    private Map<String, File> j2TemplateFiles = new HashMap<>();
-
-    private static final int TAG_KEY_COLUMN_NO_INDEX = 2;
-
-    private static final Log log = LogFactory.getLog(ConfigLoader.class);
-
-    public void validate(String migratedISHomePath, String defaultISHomePath) throws ConfigMigrateException {
-
-        if (StringUtils.isBlank(migratedISHomePath) && StringUtils.isBlank(defaultISHomePath)) {
-            throw new ConfigMigrateException("Please enter migrated IS home path and default IS home path as inputs.");
-        }
-        if (!new File(migratedISHomePath).exists() && !new File(defaultISHomePath).exists()) {
-            throw new ConfigMigrateException("The given IS home paths are not valid. Please try again!");
-        }
-        log.info("Inputs are successfully validated.");
-    }
-
-    /**
-     * Filter files by the extension and add to maps.
-     *
-     * @param migratedProductHomePath migrated IS home path.
-     * @param defaultProductHomePath default IS home path.
-     * @throws IOException IOException.
-     */
-    public void filterFiles(String migratedProductHomePath, String defaultProductHomePath) throws IOException {
-
-        defaultXMLFiles = getFilesInPath(Paths.get(defaultProductHomePath.concat(MigrationConstants.CONF_PATH)),
-                MigrationConstants.XML_FILE_EXTENSION);
-        migratedXMLFiles = getFilesInPath(Paths.get(migratedProductHomePath.concat(MigrationConstants.CONF_PATH)),
-                MigrationConstants.XML_FILE_EXTENSION);
-
-        defaultPropertiesFiles = getFilesInPath(Paths.get(defaultProductHomePath.concat(MigrationConstants.CONF_PATH)),
-                MigrationConstants.PROPERTY_FILE_EXTENSION);
-        migratedPropertiesFiles = getFilesInPath(Paths.get(migratedProductHomePath.concat(MigrationConstants
-                        .CONF_PATH)), MigrationConstants.PROPERTY_FILE_EXTENSION);
-
-        j2TemplateFiles = getFilesInPath(Paths.get(defaultProductHomePath.concat(MigrationConstants
-                .TEMPLATE_CONF_PATH)), MigrationConstants.J2_FILE_EXTENSION);
-        log.info("All files are successfully filtered by the extension.");
-    }
-
-    /**
-     * Read the knowledge from hosted  pipe separated CSV file and put that to a Map.
-     *
-     * @param csvRemoteURL Remotely hosted CSV file URL.
-     * @return Map of key and entry.
-     * @throws IOException
-     */
-     public Map<String, String> readFromCSV(URL csvRemoteURL) throws IOException {
-
-        String line;
-        Map<String, String> keys = new HashMap<>();
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(csvRemoteURL.openStream(),
-                StandardCharsets.UTF_8))) {
-            while ((line = reader.readLine()) != null) {
-                String[] parts = line.split(Pattern.quote(MigrationConstants.CSV_SEPARATOR_APPENDER));
-                if (isNewLine(parts)) {
-                    keys.put(parts[TAG_KEY_COLUMN_NO_INDEX].trim(), line);
-                }
-            }
-        }
-        return keys;
-    }
-
-    private boolean isNewLine(String[] parts) {
-
-        return parts.length > TAG_KEY_COLUMN_NO_INDEX;
-    }
-
-    /**
-     * Traverse through the files in the folder and group files using the file extension.
-     *
-     * @param defaultPath   Path to the config folder.
-     * @param fileExtension extension type to be separated.
-     * @return Map of file name and File.
-     * @throws IOException IOException
-     */
-    private Map<String, File> getFilesInPath(Path defaultPath, String fileExtension) throws IOException {
-
-        Map<String, File> mapOfFiles = new HashMap<>();
-        List<String> fileList = Files.walk(defaultPath).map(Path::toString)
-                .filter(f -> f.endsWith(fileExtension)).collect(Collectors.toList());
-        for (String path : fileList) {
-            File file = new File(path);
-            mapOfFiles.put(file.getName(), file);
-        }
-        return mapOfFiles;
-    }
-
-    public Map<String, File> getDefaultXMLFiles() {
-
-        return defaultXMLFiles;
-    }
-
-    public Map<String, File> getMigratedXMLFiles() {
-
-        return migratedXMLFiles;
-    }
-
-    public Map<String, File> getDefaultPropertiesFiles() {
-
-        return defaultPropertiesFiles;
-    }
-
-
-    public Map<String, File> getMigratedPropertiesFiles() {
-
-        return migratedPropertiesFiles;
-    }
-
-    public Map<String, File> getJ2TemplateFiles() {
-
-        return j2TemplateFiles;
-    }
-
-
-
-}
