+++ components/org.wso2.is.configuration.migrater/src/main/java/org/wso2/is/configuration/migrator/ConfigurationMigrate.java
@@ -1,0 +1,66 @@
+/*
+ *  Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
+ */
+
+package org.wso2.is.configuration.migrator;
+
+import org.wso2.is.configuration.diff.creater.DiffCheckerTool;
+import org.wso2.is.configuration.diff.creater.OutputGenerator;
+import org.wso2.is.configuration.diff.creater.exception.ConfigMigrateException;
+import org.wso2.is.configuration.migrator.util.Utils;
+import org.wso2.is.configuration.toml.generator.TomlGeneratorTool;
+
+/**
+ * Main class for the migration tool.
+ */
+public class ConfigurationMigrate {
+
+    /**
+     * The main method which runs when running the shell-script.
+     *
+     * @param args args[0] = tool type, args[1] = migrated IS home path, args[2] = default IS home path.
+     * @throws ConfigMigrateException ConfigMigrateException.
+     */
+    public static void main(String args[]) throws ConfigMigrateException {
+
+        String option = args[0];
+        String migrateISHomePath = args[1];
+        String defaultISHomePath = args[2];
+        DiffCheckerTool diffCheckTool = new DiffCheckerTool();
+        TomlGeneratorTool tomlGenTool = new TomlGeneratorTool();
+        OutputGenerator outGen;
+
+        switch (option) {
+            case "D" :
+                outGen = diffCheckTool.createDiff(migrateISHomePath, defaultISHomePath);
+                Utils.writeToFile(outGen.getKeyCatalogValuesMap(), outGen.getOutputCSV());
+                break;
+
+            case "T" :
+                outGen = diffCheckTool.createDiff(migrateISHomePath, defaultISHomePath);
+                tomlGenTool.generateTomlFile(outGen.getKeyValuesMap(), Utils.getTomlFile(defaultISHomePath),
+                        outGen.getLogFile());
+                break;
+
+            default:
+                break;
+        }
+
+    }
+
+}
