--- components/org.wso2.is.toml.generator/pom.xml
@@ -1,123 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~      http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.wso2.carbon.identity.tools</groupId>
-        <artifactId>identity-tools</artifactId>
-        <version>1.0.2-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>org.wso2.is.toml.generator</artifactId>
-    <version>1.0.2-SNAPSHOT</version>
-
-    <dependencies>
-        <dependency>
-            <groupId>commons-io.wso2</groupId>
-            <artifactId>commons-io</artifactId>
-            <version>${orbit.version.commons.io}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>commons-logging</groupId>
-            <artifactId>commons-logging</artifactId>
-            <version>${common.logging.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>commons-lang.wso2</groupId>
-            <artifactId>commons-lang</artifactId>
-            <version>${orbit.version.commons.lang}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>${commons.lang3.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.moandjiezana.toml</groupId>
-            <artifactId>toml4j</artifactId>
-            <version>${toml4j.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity.tools</groupId>
-            <artifactId>org.wso2.config.diff.creater</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <version>${mvn.assembly.plugin}</version>
-                <configuration>
-                    <descriptorRefs>
-                        <descriptorRef>jar-with-dependencies</descriptorRef>
-                    </descriptorRefs>
-                    <archive>
-                        <manifest>
-                            <mainClass>org.wso2.toml.generator.TomlGeneratorTool</mainClass>
-                        </manifest>
-                    </archive>
-                </configuration>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>8</source>
-                    <target>8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-    <properties>
-        <maven.spotbugsplugin.version>3.1.10</maven.spotbugsplugin.version>
-        <mvn.assembly.plugin>3.1.1</mvn.assembly.plugin>
-        <maven.checkstyle.plugin.version>3.0.0</maven.checkstyle.plugin.version>
-        <version.checkstyle>7.8.2</version.checkstyle>
-        <version.jinjava>2.4.14</version.jinjava>
-        <toml4j.version>0.7.2</toml4j.version>
-
-        <version.com.google.code.gson>2.8.5</version.com.google.code.gson>
-        <testng.verson>6.10</testng.verson>
-        <orbit.version.commons.io>2.4.0.wso2v1</orbit.version.commons.io>
-        <commons.lang3.version>3.9</commons.lang3.version>
-        <common.logging.version>1.2</common.logging.version>
-        <orbit.version.commons.lang>2.6.0.wso2v1</orbit.version.commons.lang>
-        <version.commons.codec>1.12</version.commons.codec>
-        <xml.unit.verions>2.6.2</xml.unit.verions>
-        <xmlunit.version>1.5</xmlunit.version>
-        <junit.version>4.12</junit.version>
-    </properties>
-
-
-</project>
